requirements.txt
ase==3.21.1
ipykernel==5.5.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tinycss2==1.1.0
pyreadline==2.1
torch-scatter==2.0.6
torch-sparse==0.6.9
torch-cluster==1.5.9
torch-spline-conv==1.2.1
torch-geometric==1.7.0
tqdm==4.60.0
scikit-learn==0.24.2
scipy==1.5.3
webencodings==0.5.1
jupyter_core==4.7.1
numpy==1.19.2
pysocks==1.7.1
tornado==6.1
setuptools==52.0.0